Added necessary classes
