+++ src/main/java/edu/wpi/N/entities/employees/EmotionalSupporter.java
@@ -1,0 +1,13 @@
+package edu.wpi.N.entities.employees;
+
+public class EmotionalSupporter extends Employee {
+
+    public EmotionalSupporter(int id, String name, String serviceType) {
+        super(id, name);
+    }
+
+    @Override
+    public String getServiceType() {
+        return "Emotional Support";
+    }
+}
